refactor(OutputView): OutputMsgContainer의 책임을 OutputView에 위임
