+++ src/main/java/baseball/controller/GameManager.java
@@ -5,20 +5,21 @@ import baseball.system.SystemConstant;
 import baseball.domain.BaseballGame;
 import baseball.view.InputView;
 import baseball.view.OutputView;
-import baseball.view.OutputMsgContainer;
 
 public class GameManager {
     private final InputView inputView = new InputView();
     private final OutputView outputView = new OutputView();
     private final BaseballGame baseballGame = new BaseballGame();
 
+    private static final String GAME_START_MESSAGE = "숫자 야구 게임을 시작합니다.\n";
+
     public void run() {
+        System.out.println(GAME_START_MESSAGE);
         startFirstGame();
         startRepeatedGame();
     }
 
     private void startFirstGame() {
-        outputView.printMessageForData(OutputMsgContainer.GAME_START_MESSAGE);
         startLoop();
     }
