+++ src/main/java/baseball/view/InputView.java
@@ -6,7 +6,10 @@ import baseball.verifier.Verifier;
 import camp.nextstep.edu.missionutils.Console;
 
 public final class InputView {
+    private final String GAME_INPUT_MESSAGE = "숫자를 입력해주세요 : ";
+
     public int readGameInput() {
+        System.out.println(GAME_INPUT_MESSAGE);
         String line = Console.readLine();
 
         Verifier baseballNumVerifier = new BaseballNumVerifier();
